Split TransformDialectStrategies into a Common, CPU and GPU parts
