+++ compiler/src/iree/compiler/Codegen/Common/TransformDialectStrategies.cpp
@@ -26,6 +26,7 @@
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/GPU/IR/GPUDialect.h"
 #include "mlir/Dialect/Linalg/IR/Linalg.h"
+#include "mlir/Dialect/Linalg/TransformOps/LinalgTransformOps.h"
 #include "mlir/Dialect/Linalg/Transforms/Transforms.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/Dialect/Tensor/IR/Tensor.h"
@@ -98,7 +99,8 @@ auto matchAndUnpack(ImplicitLocOpBuilder &b, Value targetH,
 //===----------------------------------------------------------------------===//
 
 /// Prints `handles` in order. Prints the whole IR if `handles` is empty.
-void iree_compiler::buildPrint(ImplicitLocOpBuilder &b, ValueRange handles) {
+void mlir::iree_compiler::buildPrint(ImplicitLocOpBuilder &b,
+                                     ValueRange handles) {
   if (handles.empty()) b.create<PrintOp>();
   for (auto h : handles) b.create<PrintOp>(h);
 }
@@ -124,7 +126,7 @@ void iree_compiler::buildPrint(ImplicitLocOpBuilder &b, ValueRange handles) {
 /// appended in order.
 // TODO: apply forwarding pattern.
 template <typename TilingTransformOp, typename TileOrNumThreadSpec>
-Value iree_compiler::buildTileAndFuseAndDistributeImpl(
+static Value buildTileAndFuseAndDistributeImpl(
     ImplicitLocOpBuilder &b, Value rootH, ValueRange opsHToFuse,
     ArrayRef<OpFoldResult> tileSizesOrNumThreads, ArrayAttr threadDimMapping,
     SmallVectorImpl<Value> *resultingFusedOpsHandles) {
@@ -138,9 +140,10 @@ Value iree_compiler::buildTileAndFuseAndDistributeImpl(
   return foreachThreadH;
 }
 
-/// Call buildTileAndFuseAndDistributeImpl with ArrayRef<int64_t> tilesSizes.
-template <typename TilingTransformOp = TileToForeachThreadOp>
-Value iree_compiler::buildTileFuseDistWithTileSizes(
+// TODO: if someone knows how to properly export templates go for it ..
+// sigh.
+template <typename TilingTransformOp>
+static Value buildTileFuseDistWithTileSizes(
     ImplicitLocOpBuilder &b, Value rootH, ValueRange opsHToFuse,
     ArrayRef<OpFoldResult> tileSizes, ArrayAttr threadDimMapping,
     SmallVectorImpl<Value> *resultingFusedOpsHandles) {
@@ -149,37 +152,63 @@ Value iree_compiler::buildTileFuseDistWithTileSizes(
       b, rootH, opsHToFuse, tileSizes, threadDimMapping,
       resultingFusedOpsHandles);
 }
-
-/// Call buildTileAndFuseAndDistributeImpl with ArrayRef<int64_t> numThreads.
-template <typename TilingTransformOp = TileToForeachThreadOp>
-Value iree_compiler::buildTileFuseDistWithNumThreads(
+Value mlir::iree_compiler::buildTileFuseDistToForeachThreadWithTileSizes(
     ImplicitLocOpBuilder &b, Value rootH, ValueRange opsHToFuse,
-    ArrayRef<int64_t> numThreads, ArrayAttr threadDimMapping,
+    ArrayRef<OpFoldResult> tileSizes, ArrayAttr threadDimMapping,
     SmallVectorImpl<Value> *resultingFusedOpsHandles) {
-  return buildTileAndFuseAndDistributeImpl<TilingTransformOp,
-                                           transform::NumThreadsSpec>(
-      b, rootH, opsHToFuse, getAsOpFoldResult(b.getI64ArrayAttr(numThreads)),
-      threadDimMapping, resultingFusedOpsHandles);
+  return buildTileFuseDistWithTileSizes<TileToForeachThreadOp>(
+      b, rootH, opsHToFuse, tileSizes, threadDimMapping,
+      resultingFusedOpsHandles);
+}
+Value mlir::iree_compiler::
+    buildTileFuseDistToForeachThreadAndWorgroupCountWithTileSizes(
+        ImplicitLocOpBuilder &b, Value rootH, ValueRange opsHToFuse,
+        ArrayRef<OpFoldResult> tileSizes, ArrayAttr threadDimMapping,
+        SmallVectorImpl<Value> *resultingFusedOpsHandles) {
+  return buildTileFuseDistWithTileSizes<
+      TileToForeachThreadAndWorkgroupCountRegionOp>(b, rootH, opsHToFuse,
+                                                    tileSizes, threadDimMapping,
+                                                    resultingFusedOpsHandles);
 }
 
-/// Call buildTileAndFuseAndDistributeImpl with a handle to multiple numThreads.
-template <typename TilingTransformOp = TileToForeachThreadOp>
-Value iree_compiler::buildTileFuseDistWithNumThreads(
+/// Call buildTileAndFuseAndDistributeImpl with ArrayRef<int64_t> numThreads.
+// TODO: if someone knows how to properly export templates go for it ..
+// sigh.
+template <typename TilingTransformOp>
+static Value buildTileFuseDistWithNumThreads(
     ImplicitLocOpBuilder &b, Value rootH, ValueRange opsHToFuse,
-    Value numThreads, ArrayAttr threadDimMapping,
+    ArrayRef<OpFoldResult> numThreads, ArrayAttr threadDimMapping,
     SmallVectorImpl<Value> *resultingFusedOpsHandles) {
   return buildTileAndFuseAndDistributeImpl<TilingTransformOp,
                                            transform::NumThreadsSpec>(
-      b, rootH, opsHToFuse, ArrayRef<OpFoldResult>{numThreads},
-      threadDimMapping, resultingFusedOpsHandles);
+      b, rootH, opsHToFuse, numThreads, threadDimMapping,
+      resultingFusedOpsHandles);
+}
+Value mlir::iree_compiler::buildTileFuseDistToForeachThreadWithNumThreads(
+    ImplicitLocOpBuilder &b, Value rootH, ValueRange opsHToFuse,
+    ArrayRef<OpFoldResult> tileSizes, ArrayAttr threadDimMapping,
+    SmallVectorImpl<Value> *resultingFusedOpsHandles) {
+  return buildTileFuseDistWithTileSizes<TileToForeachThreadOp>(
+      b, rootH, opsHToFuse, tileSizes, threadDimMapping,
+      resultingFusedOpsHandles);
+}
+Value mlir::iree_compiler::
+    buildTileFuseDistToForeachThreadAndWorgroupCountWithNumThreads(
+        ImplicitLocOpBuilder &b, Value rootH, ValueRange opsHToFuse,
+        ArrayRef<OpFoldResult> tileSizes, ArrayAttr threadDimMapping,
+        SmallVectorImpl<Value> *resultingFusedOpsHandles) {
+  return buildTileFuseDistWithTileSizes<
+      TileToForeachThreadAndWorkgroupCountRegionOp>(b, rootH, opsHToFuse,
+                                                    tileSizes, threadDimMapping,
+                                                    resultingFusedOpsHandles);
 }
 
 /// Apply patterns and vectorize (for now always applies rank-reduction).
 /// Takes a handle to a func.func and returns an updated handle to a
 /// func.func.
 // TODO: configure patterns.
-Value iree_compiler::buildVectorizeStrategy(ImplicitLocOpBuilder &b,
-                                            Value funcH) {
+Value mlir::iree_compiler::buildVectorizeStrategy(ImplicitLocOpBuilder &b,
+                                                  Value funcH) {
   funcH = b.create<ApplyPatternsOp>(funcH, /*rankReducing=*/true);
   return b.create<VectorizeOp>(funcH);
 }
@@ -187,9 +216,8 @@ Value iree_compiler::buildVectorizeStrategy(ImplicitLocOpBuilder &b,
 /// Post-bufferization mapping to blocks and threads.
 /// Takes a handle to a func.func and returns an updated handle to a
 /// func.func.
-Value iree_compiler::buildMapToBlockAndThreads(ImplicitLocOpBuilder &b,
-                                               Value funcH,
-                                               ArrayRef<int64_t> blockSize) {
+Value mlir::iree_compiler::buildMapToBlockAndThreads(
+    ImplicitLocOpBuilder &b, Value funcH, ArrayRef<int64_t> blockSize) {
   funcH = b.create<ForeachThreadToWorkgroupOp>(funcH);
   return b.create<MapNestedForeachThreadToGpuThreadsOp>(funcH, blockSize);
 }
@@ -199,9 +227,9 @@ static constexpr unsigned kCudaWarpSize = 32;
 /// Post-bufferization vector distribution with rank-reduction.
 /// Takes a handle to a func.func and returns an updated handle to a
 /// func.func.
-Value iree_compiler::buildDistributeVectors(ImplicitLocOpBuilder &b,
-                                            Value variantH, Value funcH,
-                                            int64_t warpSize) {
+Value mlir::iree_compiler::buildDistributeVectors(ImplicitLocOpBuilder &b,
+                                                  Value variantH, Value funcH,
+                                                  int64_t warpSize) {
   funcH = b.create<ApplyPatternsOp>(funcH, /*rankReducing=*/true);
   Value ifH = b.create<MatchOp>(funcH, scf::IfOp::getOperationName());
   // Locally suppress failures for this op only because it doesn't cover the
@@ -219,11 +247,6 @@ Value iree_compiler::buildDistributeVectors(ImplicitLocOpBuilder &b,
   return funcH;
 }
 
-//===----------------------------------------------------------------------===//
-// Higher-level problem-specific strategy creation APIs, these should favor
-// user-friendliness.
-//===----------------------------------------------------------------------===//
-
 namespace {
 /// Various handles produced by reduction splitting.
 struct ReductionSplitResult {
@@ -237,11 +260,11 @@ struct ReductionSplitResult {
   Value splitLinalgH;
   /// Handle to the final reduction.
   Value combinerH;
-  /// Handle to the original fill operation, may be null if the operation was
-  /// not re-matched.
+  /// Handle to the original fill operation, may be null if the operation
+  /// was not re-matched.
   Value originalFillH;
-  /// Handle to the trailing fill operation, may be null if the operation was
-  /// not re-matched.
+  /// Handle to the trailing fill operation, may be null if the operation
+  /// was not re-matched.
   Value trailingEltwiseH;
 };
 }  // namespace
@@ -281,11 +304,11 @@ static ReductionSplitResult createExpansionBubbleUp(
 
 /// Distribute to blocks using the current IREE lowering config.
 // TODO: consider passing a problem-specific struct to control information.
-static Value createReductionStrategyBlockDistributionPart(
+Value mlir::iree_compiler::createReductionStrategyBlockDistributionPart(
     ImplicitLocOpBuilder &b, Value variantH, Value originalFillH,
     Value reductionH, Value optionalFusionRootH,
-    ArrayRef<OpFoldResult> tileSizes0Generic, bool hasLeadingEltwise = false,
-    bool hasTrailingEltwise = false) {
+    ArrayRef<OpFoldResult> tileSizes0Generic, bool hasLeadingEltwise,
+    bool hasTrailingEltwise) {
   // Step 1. Split the reduction to get meatier parallelism.
   // TODO: use a scf.foreach_thread for this.
   auto splitReductionTransformOp =
@@ -303,8 +326,8 @@ static Value createReductionStrategyBlockDistributionPart(
   auto x = mlir::gpu::GPUBlockMappingAttr::get(b.getContext(),
                                                ::mlir::gpu::Blocks::DimX);
   // Step 2. First level of tiling + fusion parallelizes to blocks using
-  // `tileSizes`. If the fusion root was the reduction op, update it to be the
-  // combiner op. Otherwise, fuse the combiner op into root.
+  // `tileSizes`. If the fusion root was the reduction op, update it to be
+  // the combiner op. Otherwise, fuse the combiner op into root.
   SmallVector<Value> opsHToFuse(
       {rs.originalFillH ? rs.originalFillH : originalFillH, rs.splitFillH,
        rs.splitLinalgH});
@@ -319,193 +342,25 @@ static Value createReductionStrategyBlockDistributionPart(
     opsHToFuse.push_back(rs.leadingEltwiseH);
   }
 
-  // The presence of leading elementwise operation implies that dispatch region
-  // formation happened using another transform dialect script and doesn't need
-  // the workgroup count part.
+  // The presence of leading elementwise operation implies that dispatch
+  // region formation happened using another transform dialect script and
+  // doesn't need the workgroup count part.
   if (hasLeadingEltwise) {
-    iree_compiler::buildTileFuseDistWithTileSizes<TileToForeachThreadOp>(
+    iree_compiler::buildTileFuseDistToForeachThreadWithTileSizes(
         b, optionalFusionRootH, opsHToFuse, tileSizes0Generic,
         b.getArrayAttr({x}));
   } else {
-    iree_compiler::buildTileFuseDistWithTileSizes<
-        TileToForeachThreadAndWorkgroupCountRegionOp>(
-        b, optionalFusionRootH, opsHToFuse, tileSizes0Generic,
-        b.getArrayAttr({x}));
+    iree_compiler::
+        buildTileFuseDistToForeachThreadAndWorgroupCountWithTileSizes(
+            b, optionalFusionRootH, opsHToFuse, tileSizes0Generic,
+            b.getArrayAttr({x}));
   }
 
   return variantH;
 }
 
-// TODO: consider passing a problem-specific struct to control information.
-static Value createReductionStrategyThreadDistributionPart(
-    ImplicitLocOpBuilder &b, Value variantH, ArrayRef<int64_t> tileSizes1Fill,
-    ArrayRef<int64_t> tileSizes1Generic, bool hasLeadingEltwise,
-    bool hasTrailingEltwise) {
-  // TODO: Relying on ordering is brittle, harden this.
-  Value matchedH = b.create<MatchOp>(
-      variantH, ArrayRef<StringRef>{linalg::GenericOp::getOperationName(),
-                                    linalg::FillOp::getOperationName()});
-  auto split = b.create<SplitHandlesOp>(
-      matchedH,
-      /*numResultHandles=*/4 + hasLeadingEltwise + hasTrailingEltwise);
-  Value firstFusionRootH = split.getResults()[1 + hasLeadingEltwise];
-  SmallVector<Value> firstFusionGroupHs =
-      split.getResults().take_front(1 + hasLeadingEltwise);
-  Value secondFusionRootH = split.getResults().back();
-  SmallVector<Value> secondFusionGroupHs =
-      split.getResults().drop_front(2 + hasLeadingEltwise).drop_back();
-
-  auto z = mlir::gpu::GPUThreadMappingAttr::get(b.getContext(),
-                                                ::mlir::gpu::Threads::DimZ);
-  auto y = mlir::gpu::GPUThreadMappingAttr::get(b.getContext(),
-                                                ::mlir::gpu::Threads::DimY);
-
-  // clang-format off
-  iree_compiler::buildTileFuseDistWithTileSizes<TileToForeachThreadOp>(b,
-                   /*rootH=*/secondFusionRootH,
-                   /*opsHToFuse=*/secondFusionGroupHs,
-                   /*tileSizes=*/getAsOpFoldResult(b.getI64ArrayAttr(tileSizes1Fill)),
-                   /*threadDimMapping=*/b.getArrayAttr({z}));
-  iree_compiler::buildTileFuseDistWithTileSizes<TileToForeachThreadOp>(b,
-                   /*rootH=*/firstFusionRootH,
-                   /*opsHToFuse=*/firstFusionGroupHs,
-                   /*tileSizes=*/getAsOpFoldResult(b.getI64ArrayAttr(tileSizes1Generic)),
-                   /*threadDimMapping=*/b.getArrayAttr({z,y}));
-  // clang-format on
-  return variantH;
-}
-
-/// Structure to hold the parameters related to GPU reduction strategy.
-struct GPUReductionStrategyInfos {
-  std::array<int64_t, 3> workgroupSize;
-  SmallVector<int64_t> workgroupTileSizes;
-  SmallVector<int64_t> fillSecondTileSizes;
-  SmallVector<int64_t> genericSecondTileSizes;
-  bool hasLeadingEltwise;
-  bool hasTrailingEltwise;
-};
-
-/// Returns a triple of handles: the leading elementwise operation, the
-/// reduction operation and the fusion root. The leading elementwise and the
-/// fusion root may be null. If the fusion root is null, the reduction operation
-/// should be used as fusion root instead.
-// TODO: consider passing a problem-specific struct to control information.
-static std::tuple<Value, Value, Value>
-createMatchReductionBlockDistributionHandles(ImplicitLocOpBuilder &b,
-                                             Value variantH,
-                                             bool hasLeadingEltwise,
-                                             bool hasTrailingEltwise) {
-  Value originalGenericH =
-      b.create<MatchOp>(variantH, linalg::GenericOp::getOperationName());
-  auto op = b.create<SplitHandlesOp>(
-      originalGenericH,
-      /*numResultHandles=*/1 + hasLeadingEltwise + hasTrailingEltwise);
-  return std::make_tuple(hasLeadingEltwise ? op.getResults().front() : Value(),
-                         op.getResults().drop_front(hasLeadingEltwise).front(),
-                         hasTrailingEltwise ? op.getResults().back() : Value());
-}
-
-// TODO: generalize and automate over and over.
-// TODO: significantly shrink this down.
-// TODO: consider passing a problem-specific struct to control information.
-static void createReductionCudaStrategy(
-    ImplicitLocOpBuilder &b, Value variantH,
-    const GPUReductionStrategyInfos &infos) {
-  // Step 0. Match the ops.
-  Value originalFillH =
-      b.create<MatchOp>(variantH, linalg::FillOp::getOperationName());
-  auto [leadingH, reductionH, fusionRootH] =
-      createMatchReductionBlockDistributionHandles(
-          b, variantH, infos.hasLeadingEltwise, infos.hasTrailingEltwise);
-
-  // Step 1: Distribute to blocks using the current IREE lowering config.
-  variantH = createReductionStrategyBlockDistributionPart(
-      b, variantH, originalFillH, reductionH, fusionRootH,
-      getAsOpFoldResult(b.getI64ArrayAttr(infos.workgroupTileSizes)),
-      infos.hasLeadingEltwise, infos.hasTrailingEltwise);
-
-  // Step 2. Second level of tiling + fusion parallelizes to threads.
-  variantH = createReductionStrategyThreadDistributionPart(
-      b, variantH, infos.fillSecondTileSizes, infos.genericSecondTileSizes,
-      infos.hasLeadingEltwise, infos.hasTrailingEltwise);
-
-  // Step 3. Rank-reduce and vectorize.
-  // TODO: assumes a single func::FuncOp to transform, may need hardening.
-  Value funcH = b.create<MatchOp>(variantH, func::FuncOp::getOperationName());
-  funcH = iree_compiler::buildVectorizeStrategy(b, funcH);
-
-  // Step 4. Bufferize and drop HAL decriptor from memref ops.
-  variantH = b.create<IREEBufferizeOp>(variantH, /*targetGpu=*/true);
-  Value memrefFunc =
-      b.create<MatchOp>(variantH, func::FuncOp::getOperationName());
-  b.create<IREEEraseHALDescriptorTypeFromMemRefOp>(memrefFunc);
-
-  // Step 5. Post-bufferization mapping to blocks and threads.
-  // Need to match again since bufferize invalidated all handles.
-  // TODO: assumes a single func::FuncOp to transform, may need hardening.
-  funcH = b.create<MatchOp>(variantH, func::FuncOp::getOperationName());
-  funcH =
-      iree_compiler::buildMapToBlockAndThreads(b, funcH, infos.workgroupSize);
-
-  // Step 6. Post-bufferization vector distribution with rank-reduction.
-  iree_compiler::buildDistributeVectors(b, variantH, funcH);
-}
-
-// TODO: consider passing a problem-specific struct to control information.
-static bool matchGPUReduction(linalg::LinalgOp op,
-                              GPUReductionStrategyInfos &info) {
-  // TODO: match the sequence the strategy supports.
-  StructuredOpMatcher pattern, fill, leadingEltwise, trailingEltwise;
-  makeReductionMatcher(pattern, fill, leadingEltwise, trailingEltwise);
-  if (!matchPattern(op, pattern)) return false;
-
-  info.hasLeadingEltwise = leadingEltwise.getCaptured() != nullptr;
-  info.hasTrailingEltwise = trailingEltwise.getCaptured() != nullptr;
-
-  // Hardcoded workagroup size, this could be deduced from the reduction dim.
-  info.workgroupSize = {32, 2, 1};
-  SmallVector<unsigned> partitionedLoops =
-      cast<iree_compiler::PartitionableLoopsInterface>(op.getOperation())
-          .getPartitionableLoops(iree_compiler::kNumMaxParallelDims);
-  size_t numLoops = partitionedLoops.empty() ? 0 : partitionedLoops.back() + 1;
-  // Tile all the parallel dimension to 1.
-  info.workgroupTileSizes.append(numLoops, 1);
-  info.fillSecondTileSizes = {1, 0, 0};
-  info.genericSecondTileSizes = {1, 1, 0};
-  return true;
-}
-
-/// Structure to hold the parameters related to GPU reduction strategy.
-struct CPUReductionStrategyInfos {
-  int64_t workgroupSize;
-  SmallVector<int64_t> tileSizes;
-};
-
-static bool matchCPUReduction(linalg::LinalgOp op,
-                              CPUReductionStrategyInfos &infos) {
-  // TODO: match the sequence the strategy supports.
-  auto fill = m_StructuredOp<linalg::FillOp>();
-  auto pattern = m_StructuredOp()
-                     .dim(AllDims(), ShapeKind::Static)
-                     .dim(-1, utils::IteratorType::reduction)
-                     .output(NumEqualsTo(1))
-                     .output(0, fill);
-
-  // TODO: set the right config as expected by the strategy.
-  infos.workgroupSize = 1;
-  SmallVector<unsigned> partitionedLoops =
-      cast<iree_compiler::PartitionableLoopsInterface>(op.getOperation())
-          .getPartitionableLoops(iree_compiler::kNumMaxParallelDims);
-  size_t numLoops = partitionedLoops.empty() ? 0 : partitionedLoops.back() + 1;
-  // Tile all the parallel dimension to 1.
-  infos.tileSizes.append(numLoops, 1);
-  return true;
-}
-
-using StrategyBuilderFn = std::function<void(ImplicitLocOpBuilder &, Value)>;
-
-static void createTransformRegion(func::FuncOp entryPoint,
-                                  StrategyBuilderFn buildStrategy) {
+void mlir::iree_compiler::createTransformRegion(
+    func::FuncOp entryPoint, StrategyBuilderFn buildStrategy) {
   MLIRContext *ctx = entryPoint.getContext();
   Location loc = entryPoint.getLoc();
   OpBuilder b(ctx);
@@ -525,69 +380,4 @@ static void createTransformRegion(func::FuncOp entryPoint,
   LLVM_DEBUG(DBGS() << "verification: " << sequence.verify().succeeded()
                     << "\n");
   LLVM_DEBUG(sequence.print(DBGS()));
-}
-
-// TODO: generalize and automate over and over.
-// TODO: significantly shrink this down.
-static LogicalResult createReductionCpuStrategy(
-    ImplicitLocOpBuilder &b, Value variantH,
-    const CPUReductionStrategyInfos &info) {
-  // Step 0. Fetch transform information from the config and materialize it in
-  // the payload IR.
-  // TODO: this still requires specific knowledge of ops present in the IR
-  // and is very brittle.
-  Value originalFillH =
-      b.create<MatchOp>(variantH, linalg::FillOp::getOperationName());
-  Value originalGenericH =
-      b.create<MatchOp>(variantH, linalg::GenericOp::getOperationName());
-
-  // Step 1: Distribute to blocks using the current IREE lowering config.
-  variantH = createReductionStrategyBlockDistributionPart(
-      b, variantH, originalFillH, originalGenericH, Value(),
-      getAsOpFoldResult(b.getI64ArrayAttr(info.tileSizes)));
-
-  // Step 2. Rank-reduce and buildVectorizeStrategy.
-  // TODO: assumes a single func::FuncOp to transform, may need hardening.
-  Value funcH = b.create<MatchOp>(variantH, func::FuncOp::getOperationName());
-  funcH = iree_compiler::buildVectorizeStrategy(b, funcH);
-
-  // Step 3. Bufferize and drop HAL decriptor from memref ops.
-  variantH = b.create<IREEBufferizeOp>(variantH, /*targetGpu=*/true);
-  Value memrefFunc =
-      b.create<MatchOp>(variantH, func::FuncOp::getOperationName());
-  b.create<IREEEraseHALDescriptorTypeFromMemRefOp>(memrefFunc);
-
-  // Step 4. Post-bufferization mapping to blocks only.
-  // Need to match again since bufferize invalidated all handles.
-  // TODO: assumes a single func::FuncOp to transform, may need hardening.
-  funcH = b.create<MatchOp>(variantH, func::FuncOp::getOperationName());
-  funcH = b.create<ForeachThreadToWorkgroupOp>(funcH);
-
-  return success();
-}
-
-LogicalResult iree_compiler::matchAndSetGPUReductionTransformStrategy(
-    func::FuncOp entryPoint, linalg::LinalgOp op) {
-  // 1. Match
-  GPUReductionStrategyInfos infos;
-  if (!matchGPUReduction(op, infos)) return failure();
-  auto strategyBuilder = [&](ImplicitLocOpBuilder &b, Value variant) {
-    return createReductionCudaStrategy(b, variant, infos);
-  };
-  // 2. Add the strategy.
-  createTransformRegion(entryPoint, strategyBuilder);
-  return success();
-}
-
-LogicalResult iree_compiler::matchAndSetCPUReductionTransformStrategy(
-    func::FuncOp entryPoint, linalg::LinalgOp op) {
-  // 1. Match
-  CPUReductionStrategyInfos infos;
-  if (!matchCPUReduction(op, infos)) return failure();
-  auto startegyBuilder = [&](ImplicitLocOpBuilder &b, Value variant) {
-    return createReductionCpuStrategy(b, variant, infos);
-  };
-  // 2. Add the strategy.
-  createTransformRegion(entryPoint, startegyBuilder);
-  return success();
 }
